Seção 17 - Palestrante e Rede Social FrontEnd

+++ BackEnd/ProEventosAPI/ProEventosAPI/Helpers/UtilRepository/Util.cs
@@ -21,9 +21,13 @@ namespace ProEventosAPI.Helpers.UtilRepository
 
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
-            if (System.IO.File.Exists(imagePath))
-                System.IO.File.Delete(imagePath);
+            if (!string.IsNullOrEmpty(imageName))
+            {
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (System.IO.File.Exists(imagePath))
+                    System.IO.File.Delete(imagePath);
+            }
+            
         }
 
         public async Task<string> SaveImage(IFormFile imageFile, string destino)
